Move generator implementation

+++ src/board.cpp
@@ -51,7 +51,7 @@ namespace chess {
         return 1 << (rank << 3 || file);
     }
 
-    int board::toIndex(int bit) {
+    long long board::toIndex(long long bit) {
         if(bit && (bit - 1) != 0) throw std::runtime_error("Invalid Input: has multiple bits.");
         return std::__countl_zero(bit);
     }
